requirements.txt
numpy
pygetwindow
pillow